+++ yellowstone-grpc-geyser/src/grpc.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        config::{ConfigBlockFailAction, ConfigGrpc},
+        config::{ConfigBlockFailAction, ConfigGrpc, ConfigGrpcFilters},
         filters::{Filter, FilterAccountsDataSlice},
         prom::{self, DebugClientMessage, CONNECTIONS_TOTAL, MESSAGE_QUEUE_SIZE},
         version::GrpcVersionInfo,
@@ -34,9 +34,13 @@ use {
     },
     tokio_stream::wrappers::ReceiverStream,
     tonic::{
+<<<<<<< HEAD
         codec::CompressionEncoding,
         metadata::MetadataValue,
         service::{interceptor::InterceptedService, Interceptor},
+=======
+        service::interceptor::interceptor,
+>>>>>>> b9d629c2fcd570e24040bf7784757981ca9be86e
         transport::{
             server::{Server, TcpIncoming},
             Identity, ServerTlsConfig,
@@ -713,7 +717,9 @@ impl SlotMessages {
 
 #[derive(Debug)]
 pub struct GrpcService {
-    config: ConfigGrpc,
+    config_snapshot_client_channel_capacity: usize,
+    config_channel_capacity: usize,
+    config_filters: Arc<ConfigGrpcFilters>,
     blocks_meta: Option<BlockMetaStorage>,
     subscribe_id: AtomicUsize,
     snapshot_rx: Mutex<Option<crossbeam_channel::Receiver<Option<Message>>>>,
@@ -777,18 +783,32 @@ impl GrpcService {
 
         // Create Server
         let max_decoding_message_size = config.max_decoding_message_size;
+<<<<<<< HEAD
         let service = GeyserServer::new(Self {
             config: config.clone(),
+=======
+        let mut service = GeyserServer::new(Self {
+            config_snapshot_client_channel_capacity: config.snapshot_client_channel_capacity,
+            config_channel_capacity: config.channel_capacity,
+            config_filters: Arc::new(config.filters),
+>>>>>>> b9d629c2fcd570e24040bf7784757981ca9be86e
             blocks_meta,
             subscribe_id: AtomicUsize::new(0),
             snapshot_rx: Mutex::new(snapshot_rx),
             broadcast_tx: broadcast_tx.clone(),
             debug_clients_tx,
         })
-        .accept_compressed(CompressionEncoding::Gzip)
-        .send_compressed(CompressionEncoding::Gzip)
         .max_decoding_message_size(max_decoding_message_size);
+<<<<<<< HEAD
         let service = InterceptedService::new(service, XTokenChecker::new(config.x_token));
+=======
+        for encoding in config.compression.accept {
+            service = service.accept_compressed(encoding);
+        }
+        for encoding in config.compression.send {
+            service = service.send_compressed(encoding);
+        }
+>>>>>>> b9d629c2fcd570e24040bf7784757981ca9be86e
 
         // Run geyser message loop
         let (messages_tx, messages_rx) = mpsc::unbounded_channel();
@@ -817,6 +837,16 @@ impl GrpcService {
 
             server_builder
                 .http2_keepalive_interval(Some(Duration::from_secs(5)))
+                .layer(interceptor(move |request: Request<()>| {
+                    if let Some(x_token) = &config.x_token {
+                        match request.metadata().get("x-token") {
+                            Some(token) if x_token == token => Ok(request),
+                            _ => Err(Status::unauthenticated("No valid auth token")),
+                        }
+                    } else {
+                        Ok(request)
+                    }
+                }))
                 .add_service(health_service)
                 .add_service(service)
                 .serve_with_incoming_shutdown(incoming, shutdown_grpc.notified())
@@ -1094,7 +1124,8 @@ impl GrpcService {
     #[allow(clippy::too_many_arguments)]
     async fn client_loop(
         id: usize,
-        mut filter: Filter,
+        endpoint: String,
+        config_filters: Arc<ConfigGrpcFilters>,
         stream_tx: mpsc::Sender<TonicResult<SubscribeUpdate>>,
         mut client_rx: mpsc::UnboundedReceiver<Option<Filter>>,
         mut snapshot_rx: Option<crossbeam_channel::Receiver<Option<Message>>>,
@@ -1102,6 +1133,24 @@ impl GrpcService {
         debug_client_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
         drop_client: impl FnOnce(),
     ) {
+        let mut filter = Filter::new(
+            &SubscribeRequest {
+                accounts: HashMap::new(),
+                slots: HashMap::new(),
+                transactions: HashMap::new(),
+                transactions_status: HashMap::new(),
+                blocks: HashMap::new(),
+                blocks_meta: HashMap::new(),
+                entry: HashMap::new(),
+                commitment: None,
+                accounts_data_slice: Vec::new(),
+                ping: None,
+            },
+            &config_filters,
+        )
+        .expect("empty filter");
+        prom::update_subscriptions(&endpoint, None, Some(&filter));
+
         CONNECTIONS_TOTAL.inc();
         DebugClientMessage::maybe_send(&debug_client_tx, || DebugClientMessage::UpdateFilter {
             id,
@@ -1126,6 +1175,7 @@ impl GrpcService {
                             continue;
                         }
 
+                        prom::update_subscriptions(&endpoint, Some(&filter), Some(&filter_new));
                         filter = filter_new;
                         info!("client #{id}: filter updated");
                     }
@@ -1182,6 +1232,7 @@ impl GrpcService {
                                     continue;
                                 }
 
+                                prom::update_subscriptions(&endpoint, Some(&filter), Some(&filter_new));
                                 filter = filter_new;
                                 DebugClientMessage::maybe_send(&debug_client_tx, || DebugClientMessage::UpdateFilter { id, filter: Box::new(filter.clone()) });
                                 info!("client #{id}: filter updated");
@@ -1244,6 +1295,7 @@ impl GrpcService {
 
         CONNECTIONS_TOTAL.dec();
         DebugClientMessage::maybe_send(&debug_client_tx, || DebugClientMessage::Removed { id });
+        prom::update_subscriptions(&endpoint, Some(&filter), None);
         info!("client #{id}: removed");
         drop_client();
     }
@@ -1258,27 +1310,11 @@ impl Geyser for GrpcService {
         mut request: Request<Streaming<SubscribeRequest>>,
     ) -> TonicResult<Response<Self::SubscribeStream>> {
         let id = self.subscribe_id.fetch_add(1, Ordering::Relaxed);
-        let filter = Filter::new(
-            &SubscribeRequest {
-                accounts: HashMap::new(),
-                slots: HashMap::new(),
-                transactions: HashMap::new(),
-                transactions_status: HashMap::new(),
-                blocks: HashMap::new(),
-                blocks_meta: HashMap::new(),
-                entry: HashMap::new(),
-                commitment: None,
-                accounts_data_slice: Vec::new(),
-                ping: None,
-            },
-            &self.config.filters,
-        )
-        .expect("empty filter");
         let snapshot_rx = self.snapshot_rx.lock().await.take();
         let (stream_tx, stream_rx) = mpsc::channel(if snapshot_rx.is_some() {
-            self.config.snapshot_client_channel_capacity
+            self.config_snapshot_client_channel_capacity
         } else {
-            self.config.channel_capacity
+            self.config_channel_capacity
         });
         let (client_tx, client_rx) = mpsc::unbounded_channel();
         let notify_exit1 = Arc::new(Notify::new());
@@ -1315,7 +1351,13 @@ impl Geyser for GrpcService {
             }
         });
 
-        let config_filters_limit = self.config.filters.clone();
+        let endpoint = request
+            .metadata()
+            .get("x-endpoint")
+            .and_then(|h| h.to_str().ok().map(|s| s.to_string()))
+            .unwrap_or_else(|| "".to_owned());
+
+        let config_filters = Arc::clone(&self.config_filters);
         let incoming_stream_tx = stream_tx.clone();
         let incoming_client_tx = client_tx;
         let incoming_exit = Arc::clone(&notify_exit2);
@@ -1330,7 +1372,7 @@ impl Geyser for GrpcService {
                     }
                     message = request.get_mut().message() => match message {
                         Ok(Some(request)) => {
-                            if let Err(error) = match Filter::new(&request, &config_filters_limit) {
+                            if let Err(error) = match Filter::new(&request, &config_filters) {
                                 Ok(filter) => match incoming_client_tx.send(Some(filter)) {
                                     Ok(()) => Ok(()),
                                     Err(error) => Err(error.to_string()),
@@ -1359,7 +1401,8 @@ impl Geyser for GrpcService {
 
         tokio::spawn(Self::client_loop(
             id,
-            filter,
+            endpoint,
+            Arc::clone(&self.config_filters),
             stream_tx,
             client_rx,
             snapshot_rx,
@@ -1388,13 +1431,14 @@ impl Geyser for GrpcService {
             blocks_meta
                 .get_block(
                     |block| {
-                        block.block_height.map(|last_valid_block_height| {
-                            GetLatestBlockhashResponse {
+                        block
+                            .block_height
+                            .map(|block_height| GetLatestBlockhashResponse {
                                 slot: block.slot,
                                 blockhash: block.blockhash.clone(),
-                                last_valid_block_height,
-                            }
-                        })
+                                last_valid_block_height: block_height
+                                    + MAX_RECENT_BLOCKHASHES as u64,
+                            })
                     },
                     request.get_ref().commitment,
                 )